requirements.txt
asgiref==3.3.4
Django==3.2
django-appconf==1.0.4
django-cleanup==5.2.0
django-image-cropping==1.5.0
easy-thumbnails==2.7.1
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
transliterate==1.10.2